Put the new code in a better spot
